Disable IPv6 in Google Compute Engine
Not currently supported in the Google Compute Engine network, so except
for esoteric use cases there's no need to run the risk of software
handling this badly. With support in the kernel but missing in the
network, we've already seen one case of misleading error behavior that,
while not technically incorrect, made debugging more difficult.