Remove spurious use of kpartx
A previous internal iteration of commit
a79d2e490671e8aa465b4215a9e570d3d3f6af69 used kpartx, but the final version
didn't. Remove the logic in the cleanup code which tries to tear down kpartx.
Ironically this would fail on a system without kpartx installed, and it's not
needed as a dependency in the final version.